+++ include/aca_ovs_control.h
@@ -19,21 +19,22 @@
 #define DSCP_DEFAULT (IPTOS_PREC_INTERNETCONTROL >> 2)
 #define STDOUT_FILENO   1   /* Standard output.  */
 
-#include "goalstateprovisioner.grpc.pb.h"
-#include <openvswitch/vconn-provider.h> /* add to /usr/local/include/openvswitch */
-#include <openvswitch/ofpbuf.h>
+// #include "goalstateprovisioner.grpc.pb.h"
+// #include <openvswitch/vconn-provider.h> /* add to /usr/local/include/openvswitch */
+// #include <openvswitch/ofpbuf.h>
 #include <openvswitch/ofp-errors.h>
 #include <openvswitch/ofp-packet.h>
-#include <openvswitch/ofp-port.h>
-#include <openvswitch/ofp-table.h>
-#include <openvswitch/ofp-protocol.h>
-#include <openvswitch/ofp-switch.h>
-#include <linux/in.h>
+// #include <openvswitch/ofp-port.h>
+// #include <openvswitch/ofp-table.h>
+// #include <openvswitch/ofp-protocol.h>
+// #include <openvswitch/ofp-switch.h>
+// #include <linux/in.h>
+// #include <linux/types.h>
 #include <string>
 
-extern "C" { 
-    struct unixctl_conn;
-}
+// extern "C" { 
+//     struct unixctl_conn;
+// }
 // OVS monitor implementation class
 namespace aca_ovs_control
 {
@@ -41,126 +42,133 @@ class ACA_OVS_Control {
   public:
   static ACA_OVS_Control &get_instance();
 
-/* --names, --no-names: Show port and table names in output and accept them in
- * input.  (When neither is specified, the default is to accept port names but,
- * for backward compatibility, not to show them unless this is an interactive
- * console session.)  */
-  static int use_names;
-  static int verbosity;
-  /* -F, --flow-format: Allowed protocols.  By default, any protocol is allowed. */
-  static enum ofputil_protocol allowed_protocols;
-  /* --unixctl-path: Path to use for unixctl server, for "monitor" and "snoop"
-     commands. */
-  char *unixctl_path;
+  /* --names, --no-names: Show port and table names in output and accept them in
+   * input.  (When neither is specified, the default is to accept port names but,
+   * for backward compatibility, not to show them unless this is an interactive
+   * console session.)  */
+//   static int use_names;
+//   static int verbosity;
+//   /* -F, --flow-format: Allowed protocols.  By default, any protocol is allowed. */
+//   static enum ofputil_protocol allowed_protocols;
+//   /* --unixctl-path: Path to use for unixctl server, for "monitor" and "snoop"
+//      commands. */
+//   char *unixctl_path;
 
   int control();
-  int get_packet_in(const char *bridge); 
+  int monitor(const char *bridge, const char *opt);
+  int packet_out(const char *bridge, const char *opt);  
+  int dump_flows(const char *bridge, const char *opt);
+  //int get_packet_in(const char *bridge); 
   void parse_packet(void *packet);
   void print_payload(const u_char *payload, int len);
   void print_hex_ascii_line(const u_char *payload, int len, int offset);
-  int monitor(const char *bridge, const char *opt);
-  void packet_out(const char *bridge, const char *opt);
+
   
   /* 
    * structs and funcntions borrow from ovs-ofctl.c 
    */
-  void dump_flows(const char *bridge, const char *opt);
-  void dump_flows__(int argc, const char *argv, bool aggregate);
-  vconn *prepare_dump_flows(int argc, const char *argv, bool aggregate,
-                    ofputil_flow_stats_request *fsr,
-                    ofputil_protocol *protocolp);
-  enum ofputil_protocol set_protocol_for_flow_dump(vconn *vconn,
-                           ofputil_protocol cur_protocol,
-                           ofputil_protocol usable_protocols);
-  enum ofputil_protocol open_vconn_for_flow_mod(const char *remote, vconn **vconnp,
-                        ofputil_protocol usable_protocols);
-  bool try_set_protocol(struct vconn *vconn, enum ofputil_protocol want,
-                 enum ofputil_protocol *cur);
-  void fetch_switch_config(vconn *vconn, ofputil_switch_config *config);
-  void set_switch_config(vconn *vconn, const ofputil_switch_config *config);              
-  int open_vconn_socket(const char *name, vconn **vconnp);
-  void run(int retval, const char *message, ...);
-  enum ofputil_protocol open_vconn(const char *name, vconn **vconnp);
-  void transact_noreply(vconn *vconn, ofpbuf *request);
-  void transact_multiple_noreply(vconn *vconn, ovs_list *requests);
-  int monitor_set_invalid_ttl_to_controller(vconn *vconn);
-  bool set_packet_in_format(vconn *vconn,
-                     enum ofputil_packet_in_format packet_in_format,
-                     bool must_succeed);
-  void monitor_vconn(vconn *vconn, bool reply_to_echo_requests,
-               bool resume_continuations, const char *bridge);
-  void send_openflow_buffer(vconn *vconn, ofpbuf *buffer);
-  void dump_transaction(vconn *vconn, ofpbuf *request, const char *bridge);
+//   void dump_flows(const char *bridge, const char *opt);
+//   void dump_flows__(int argc, const char *argv, bool aggregate);
+//   vconn *prepare_dump_flows(int argc, const char *argv, bool aggregate,
+//                     ofputil_flow_stats_request *fsr,
+//                     ofputil_protocol *protocolp);
+//   enum ofputil_protocol set_protocol_for_flow_dump(vconn *vconn,
+//                            ofputil_protocol cur_protocol,
+//                            ofputil_protocol usable_protocols);
+//   enum ofputil_protocol open_vconn_for_flow_mod(const char *remote, vconn **vconnp,
+//                         ofputil_protocol usable_protocols);
+//   bool try_set_protocol(struct vconn *vconn, enum ofputil_protocol want,
+//                  enum ofputil_protocol *cur);
+//   void fetch_switch_config(vconn *vconn, ofputil_switch_config *config);
+//   void set_switch_config(vconn *vconn, const ofputil_switch_config *config);              
+//   int open_vconn_socket(const char *name, vconn **vconnp);
+//   void run(int retval, const char *message, ...);
+//   enum ofputil_protocol open_vconn(const char *name, vconn **vconnp);
+//   void transact_noreply(vconn *vconn, ofpbuf *request);
+//   void transact_multiple_noreply(vconn *vconn, ovs_list *requests);
+//   int monitor_set_invalid_ttl_to_controller(vconn *vconn);
+//   bool set_packet_in_format(vconn *vconn,
+//                      enum ofputil_packet_in_format packet_in_format,
+//                      bool must_succeed);
+//   void monitor_vconn(vconn *vconn, bool reply_to_echo_requests,
+//                bool resume_continuations, const char *bridge);
+//   void send_openflow_buffer(vconn *vconn, ofpbuf *buffer);
+//   void dump_transaction(vconn *vconn, ofpbuf *request, const char *bridge);
   
-  struct barrier_aux {
-    struct vconn *vconn;        /* OpenFlow connection for sending barrier. */
-    struct unixctl_conn *conn;  /* Connection waiting for barrier response. */
-  };
+//   struct barrier_aux {
+//     struct vconn *vconn;        /* OpenFlow connection for sending barrier. */
+//     struct unixctl_conn *conn;  /* Connection waiting for barrier response. */
+//   };
 
-  enum PI { PI_FEATURES, PI_PORT_DESC }; 
-  struct port_iterator {
-      struct vconn *vconn;
-      PI variant;
-      struct ofpbuf *reply;
-      ovs_be32 send_xid;
-      bool more;
-  };
-  enum TI { TI_STATS, TI_FEATURES };
-  struct table_iterator {
-    struct vconn *vconn;
-    TI variant;
-    struct ofpbuf *reply;
-    ovs_be32 send_xid;
-    bool more;
+//   enum PI { PI_FEATURES, PI_PORT_DESC }; 
+//   struct port_iterator {
+//       struct vconn *vconn;
+//       PI variant;
+//       struct ofpbuf *reply;
+//       ovs_be32 send_xid;
+//       bool more;
+//   };
+//   enum TI { TI_STATS, TI_FEATURES };
+//   struct table_iterator {
+//     struct vconn *vconn;
+//     TI variant;
+//     struct ofpbuf *reply;
+//     ovs_be32 send_xid;
+//     bool more;
 
-    struct ofputil_table_features features;
-    struct ofpbuf raw_properties;
-  };
+//     struct ofputil_table_features features;
+//     struct ofpbuf raw_properties;
+//   };
 
-  ofp_port_t str_to_port_no(const char *vconn_name, const char *port_name);
-  bool str_to_ofp(const char *s, ofp_port_t *ofp_port);
-  void port_iterator_fetch_port_desc(port_iterator *pi);
-  void port_iterator_fetch_features(port_iterator *pi);
-  void port_iterator_init(port_iterator *pi, vconn *vconn);
-  bool port_iterator_next(port_iterator *pi, ofputil_phy_port *pp);
-  void port_iterator_destroy(port_iterator *pi);
-  void fetch_ofputil_phy_port(const char *vconn_name, const char *port_name, ofputil_phy_port *pp);
-  const ofputil_port_map *get_port_map(const char *vconn_name);
-  const ofputil_port_map *ports_to_accept(const char *vconn_name);
-  const ofputil_port_map *ports_to_show(const char *vconn_name);
-  void table_iterator_init(table_iterator *ti, struct vconn *vconn);
-  const ofputil_table_features * table_iterator_next(table_iterator *ti);
-  void table_iterator_destroy(table_iterator *ti);
-  const ofputil_table_map *get_table_map(const char *vconn_name);
-  const ofputil_table_map *tables_to_accept(const char *vconn_name);
-  const ofputil_table_map *tables_to_show(const char *vconn_name);
-  bool should_accept_names(void);
-  bool should_show_names(void);
-  const char * openflow_from_hex(const char *hex, ofpbuf **msgp);
+//   ofp_port_t str_to_port_no(const char *vconn_name, const char *port_name);
+//   bool str_to_ofp(const char *s, ofp_port_t *ofp_port);
+//   void port_iterator_fetch_port_desc(port_iterator *pi);
+//   void port_iterator_fetch_features(port_iterator *pi);
+//   void port_iterator_init(port_iterator *pi, vconn *vconn);
+//   bool port_iterator_next(port_iterator *pi, ofputil_phy_port *pp);
+//   void port_iterator_destroy(port_iterator *pi);
+//   void fetch_ofputil_phy_port(const char *vconn_name, const char *port_name, ofputil_phy_port *pp);
+//   const ofputil_port_map *get_port_map(const char *vconn_name);
+//   const ofputil_port_map *ports_to_accept(const char *vconn_name);
+//   const ofputil_port_map *ports_to_show(const char *vconn_name);
+//   void table_iterator_init(table_iterator *ti, struct vconn *vconn);
+//   const ofputil_table_features * table_iterator_next(table_iterator *ti);
+//   void table_iterator_destroy(table_iterator *ti);
+//   const ofputil_table_map *get_table_map(const char *vconn_name);
+//   const ofputil_table_map *tables_to_accept(const char *vconn_name);
+//   const ofputil_table_map *tables_to_show(const char *vconn_name);
+//   bool should_accept_names(void);
+//   bool should_show_names(void);
+//   const char * openflow_from_hex(const char *hex, ofpbuf **msgp);
 
   /* ethernet headers are always exactly 14 bytes [1] */
   #define SIZE_ETHERNET 14
 
+  /* Internet address. */
+//   struct in_addr {
+//           __u32	s_addr;
+//   };
+
   /* IP header */
-  // struct sniff_ip {
-  //         u_char  ip_vhl;                 /* version << 4 | header length >> 2 */
-  //         u_char  ip_tos;                 /* type of service */
-  //         u_short ip_len;                 /* total length */
-  //         u_short ip_id;                  /* identification */
-  //         u_short ip_off;                 /* fragment offset field */
-  //         #define IP_RF 0x8000            /* reserved fragment flag */
-  //         #define IP_DF 0x4000            /* dont fragment flag */
-  //         #define IP_MF 0x2000            /* more fragments flag */
-  //         #define IP_OFFMASK 0x1fff       /* mask for fragmenting bits */
-  //         u_char  ip_ttl;                 /* time to live */
-  //         u_char  ip_p;                   /* protocol */
-  //         u_short ip_sum;                 /* checksum */
-  //         struct  in_addr ip_src,ip_dst;  /* source and dest address */
-  // };
+  struct sniff_ip {
+          u_char  ip_vhl;                 /* version << 4 | header length >> 2 */
+          u_char  ip_tos;                 /* type of service */
+          u_short ip_len;                 /* total length */
+          u_short ip_id;                  /* identification */
+          u_short ip_off;                 /* fragment offset field */
+          #define IP_RF 0x8000            /* reserved fragment flag */
+          #define IP_DF 0x4000            /* dont fragment flag */
+          #define IP_MF 0x2000            /* more fragments flag */
+          #define IP_OFFMASK 0x1fff       /* mask for fragmenting bits */
+          u_char  ip_ttl;                 /* time to live */
+          u_char  ip_p;                   /* protocol */
+          u_short ip_sum;                 /* checksum */
+          struct in_addr ip_src, ip_dst;  /* source and dest address */
+  };
   /* The second-half of the first byte in ip_header contains the IP header length (IHL). */
-  // #define IP_HL(ip)               (((ip)->ip_vhl) & 0x0f)
-  // /* The IHL is number of 32-bit segments. Multiply by four to get a byte count for pointer arithmetic */
-  // #define IP_V(ip)                (((ip)->ip_vhl) >> 4)
+  #define IP_HL(ip)               (((ip)->ip_vhl) & 0x0f)
+  /* The IHL is number of 32-bit segments. Multiply by four to get a byte count for pointer arithmetic */
+  #define IP_V(ip)                (((ip)->ip_vhl) >> 4)
 
   /* TCP header */
   typedef u_int tcp_seq;
@@ -185,6 +193,14 @@ class ACA_OVS_Control {
           u_short th_win;                 /* window */
           u_short th_sum;                 /* checksum */
           u_short th_urp;                 /* urgent pointer */
+  };
+
+  /* UDP protocol header. */
+  struct sniff_udp {
+          u_short uh_sport;               /* source port */
+          u_short uh_dport;               /* destination port */
+          u_short uh_ulen;                /* udp length */
+          u_short uh_sum;                 /* udp checksum */
   };
 
   // compiler will flag the error when below is called.
